pipeline/requirements.txt
geopandas>=1.0
psycopg2-binary>=2.9
sqlalchemy>=2.0
python-dotenv>=1.0
httpx>=0.27
pytest>=8.0
